pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "azure-rag-app"
version = "0.1.0"
description = "RAG app using Azure OpenAI, Azure AI Search, FastAPI, vanilla JS"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "openai>=1.0.0",
    "azure-search-documents>=11.6.0",
    "azure-core>=1.30.0",
    "jinja2",
    "pydantic>=2.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
